コントローラー　修正

+++ app/controllers/items_controller.rb
@@ -1,6 +1,6 @@
 class ItemsController < ApplicationController
   before_action :authenticate_user!, only: [:new, :edit, :destroy]
-  before_action :set_item,      only: [:show,:edit,:update]
+  before_action :set_item,      only: [:show,:edit,:update,:destroy]
   before_action :move_to_index,      only: [:edit]
 
   def index
@@ -35,9 +35,13 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-    item = Item.find(params[:id])
-    item.destroy
-    redirect_to root_path
+    if current_user.id == @item.user_id
+      item = Item.find(params[:id])
+      item.destroy
+      redirect_to root_path
+    else
+      render :index
+    end
   end
 
   private
